refactor (model): Uses relatedlistTrait
